Spostati metodi legati agli e-service descriptor
