+++ interop-qa-tests/src/main/java/it/pagopa/interop/eservice/service/IM2MEserviceDescriptorClient.java
@@ -3,7 +3,9 @@ package it.pagopa.interop.eservice.service;
 import it.pagopa.interop.agreement.domain.EServiceDescriptor;
 import it.pagopa.interop.common.client.IClient;
 import it.pagopa.interop.generated.openapi.clients.m2mGateway.model.AgreementApprovalPolicy;
+import it.pagopa.interop.generated.openapi.clients.m2mGateway.model.Documents;
 import it.pagopa.interop.generated.openapi.clients.m2mGateway.model.EServiceDescriptorState;
+import it.pagopa.interop.generated.openapi.clients.m2mGateway.model.FileDownloadMultipart;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NonNull;
@@ -43,9 +45,14 @@ public interface IM2MEserviceDescriptorClient extends IClient<EServiceDescriptor
     }
 
     EServiceDescriptor get(UUID eserviceId, UUID descriptorId);
+    it.pagopa.interop.generated.openapi.clients.m2mGateway.model.EServiceDescriptor getDescriptor(UUID eserviceId, UUID descriptorId);
     it.pagopa.interop.generated.openapi.clients.m2mGateway.model.EServiceDescriptor getCompleteResource(UUID eserviceId, UUID descriptorId);
     List<EServiceDescriptor> getAll(EserviceDescriptorsListRequest eserviceDescriptorsListRequest);
     List<EServiceDescriptor> getAll(UUID eserviceId);
+    void deleteInterface(UUID eServiceId, UUID descriptorId);
+    void unsuspendEService(UUID eServiceId, UUID descriptorId);
+    FileDownloadMultipart downloadEServiceDescriptorInterface(UUID eserviceId, UUID descriptorId);
+    Documents getDocuments(UUID eserviceId, UUID descriptorId);
     it.pagopa.interop.generated.openapi.clients.m2mGateway.model.EServiceDescriptor patchEServiceDescriptor(UUID eserviceId, UUID descriptorId, EServiceDescriptorPatchRequest body);
     it.pagopa.interop.generated.openapi.clients.m2mGateway.model.EServiceDescriptor patchEServiceDescriptorQuotas(UUID eserviceId, UUID descriptorId, EServiceDescriptorQuotasPatchRequest body);
 }
